Move styles to app.js to theme whole app
