+++ ronaapp/src/components/SignUp.js
@@ -7,18 +7,7 @@ import { FormControl, InputLabel, Input, FormHelperText, ButtonGroup, Button } f
 import Grid from '@material-ui/core/Grid';
 import Container from '@material-ui/core/Container';
 import Box from '@material-ui/core/Box';
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 
-const theme = createMuiTheme({
-  palette: {
-    primary:{
-        main: '#00695c'
-    },
-    secondary: {
-      main: '#E33E7F'
-    }
-  }
-});
 
 function SignUp() {
     const { currentUser } = useContext(AuthContext);
@@ -49,50 +38,49 @@ function SignUp() {
 
     return (
         <div>
-            <MuiThemeProvider theme={theme}>
 
-            <Container component="main" maxWidth="xs">
+                <Container component="main" maxWidth="xs">
 
-                <h1>Sign up</h1>
+                    <h1>Sign up</h1>
 
-                <form onSubmit={handleSignUp}>
+                    <form onSubmit={handleSignUp}>
 
-                    <Box mt={4}>
-                        <FormControl fullWidth>
-                            <InputLabel htmlFor="name-input">Name</InputLabel>
-                            <Input id="name-input" name="displayName" required="true" autoFocus="true" />
-                        </FormControl>
-                    </Box>
+                        <Box mt={4}>
+                            <FormControl fullWidth>
+                                <InputLabel htmlFor="name-input">Name</InputLabel>
+                                <Input id="name-input" name="displayName" required="true" autoFocus="true" />
+                            </FormControl>
+                        </Box>
 
 
-                    <Box mt={4}>
-                        <FormControl fullWidth>
-                            <InputLabel htmlFor="email-input">Email</InputLabel>
-                            <Input id="email-input" name="email" type="email" required="true" />
-                        </FormControl>
-                    </Box>
+                        <Box mt={4}>
+                            <FormControl fullWidth>
+                                <InputLabel htmlFor="email-input">Email</InputLabel>
+                                <Input id="email-input" name="email" type="email" required="true" />
+                            </FormControl>
+                        </Box>
 
-                    <Box mt={4}>
-                        <FormControl >
-                            <InputLabel htmlFor="passwordOne-input">Password</InputLabel>
-                            <Input id="passwordOne-input" name="passwordOne" type="password" required="true" />
-                        </FormControl>
+                        <Box mt={4}>
+                            <FormControl >
+                                <InputLabel htmlFor="passwordOne-input">Password</InputLabel>
+                                <Input id="passwordOne-input" name="passwordOne" type="password" required="true" />
+                            </FormControl>
 
-                        <FormControl >
-                            <InputLabel htmlFor="passwordTwo-input">Confirm Password</InputLabel>
-                            <Input id="passwordTwo-input" name="passwordTwo" type="password" required="true" />
-                            {pwMatch && <FormHelperText error className="error" id="password-helper-text">{pwMatch}</FormHelperText>}
+                            <FormControl >
+                                <InputLabel htmlFor="passwordTwo-input">Confirm Password</InputLabel>
+                                <Input id="passwordTwo-input" name="passwordTwo" type="password" required="true" />
+                                {pwMatch && <FormHelperText error className="error" id="password-helper-text">{pwMatch}</FormHelperText>}
 
-                        </FormControl>
-                    </Box>
+                            </FormControl>
+                        </Box>
 
-                    <Box mt={4}>
-                        <Button type="submit" variant="contained" color="primary" fullWidth>
-                            Log in
-                        </Button>
-                    </Box>
+                        <Box mt={4}>
+                            <Button type="submit" variant="contained" color="primary" fullWidth>
+                                Log in
+                            </Button>
+                        </Box>
 
-                    {/* <div className="form-group">
+                        {/* <div className="form-group">
                     <label>
                         Name:
                         <input
@@ -144,10 +132,9 @@ function SignUp() {
                 <Button type="submit" variant="contained" color="primary">
                     Sign in
                 </Button> */}
-                </form>
-                <SocialSignIn />
-            </Container>
-            </MuiThemeProvider>
+                    </form>
+                    <SocialSignIn />
+                </Container>
 
         </div>
     );
